Observer data, ScaleSynchronizer added to sync multiple graphs scale

+++ AviaryStructural/src/TimeGraph.java
@@ -316,15 +316,15 @@ public class TimeGraph {
 
         App.processingRef.fill(this.valueTextCl.getRGB());
         App.processingRef.textSize(this.textSize);
-        App.processingRef.text((int)(this.dots.get(this.dots.size() - 1).getY()), this.originX + this.textSize + 40, (float)(lastAbsoluteCoordinates.getY() - (this.textSize - 4)));
+        App.processingRef.text((float)(this.dots.get(this.dots.size() - 1).getY()), this.originX + this.textSize + 40, (float)(lastAbsoluteCoordinates.getY() - (this.textSize - 4)));
     }
 
     void renderAxisScale() {
         App.processingRef.fill(this.valueTextCl.getRGB());
         App.processingRef.textSize(this.textSize);
         App.processingRef.text(App.processingRef.millis()/1000, this.originX + this.dimX - (this.textSize + 20), this.originY + this.dimY - (this.textSize - 4));
-        App.processingRef.text((int)(this.maxY + 0.25 * (this.maxY - this.minY)), this.originX + 5, this.originY + this.textSize + 4);
-        App.processingRef.text((int)(this.minY), this.originX + 5, this.originY + this.dimY - (this.textSize - 4));
+        App.processingRef.text((float)(this.maxY + 0.25 * (this.maxY - this.minY)), this.originX + 5, this.originY + this.textSize + 4);
+        App.processingRef.text((float)(this.minY), this.originX + 5, this.originY + this.dimY - (this.textSize - 4));
     }
 
     void renderTitle() {
